Refactoring UserService and AuthenticationCredentialsService
