+++ app/application/uses_cases/auth.py
@@ -1,6 +1,6 @@
 from app.core.exception import UnauthorizedAccessException
 from app.core.types import Token, UserCode
-from app.domain.services.user import AuthenticationCredentialsService
+from app.domain.services.auth import AuthenticationCredentialsService
 from app.infrastructure.repository.user import UserRepository
 from app.presentation.schemes import UserResponse
 from app.core.data import get_username
